pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "pixel-art-pipeline"
version = "0.1.0"
description = "Batch pixel art animation generator using PixelLab API"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [
    {name = "Ormus", email = "hermeticormus@gmail.com"},
]
keywords = ["pixel-art", "animation", "pixellab", "sprite", "gif"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Multimedia :: Graphics",
]
dependencies = [
    "pillow>=10.0",
    "requests>=2.28",
    "pyyaml>=6.0",
]

[project.urls]
Homepage = "https://github.com/HermeticOrmus/pixel-art-pipeline"
Repository = "https://github.com/HermeticOrmus/pixel-art-pipeline"
Issues = "https://github.com/HermeticOrmus/pixel-art-pipeline/issues"

[project.scripts]
pixelart = "pixelart.cli:main"

[tool.ruff]
target-version = "py310"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "W", "I"]

[tool.pytest.ini_options]
testpaths = ["tests"]
